修复了各种事件的序列化异常

+++ RhythmBaseCS/RhythmDoctor/Events/SetHeartExplodeInterval.cs
@@ -20,7 +20,7 @@
 		/// <summary>
 		/// Gets or sets the interval value.
 		/// </summary>
-		public int Interval { get; set; }
+		public float Interval { get; set; }
 		/// <summary>
 		/// Gets the type of the event.
 		/// </summary>
